Move FileManager out of editing module
