+++ indico/modules/events/papers/client/js/components/PaperSubmissionButton.tsx
@@ -15,7 +15,7 @@ import _ from 'lodash';
 import React, {useState} from 'react';
 import {Button} from 'semantic-ui-react';
 
-import {FinalFileManager} from 'indico/modules/events/editing/editing/timeline/FileManager';
+import {FinalFileManager} from 'indico/react/components/files/FileManager';
 import {handleSubmitError} from 'indico/react/forms';
 import {FinalModalForm} from 'indico/react/forms/final-form';
 import {useIndicoAxios} from 'indico/react/hooks';
