+++ indico/web/client/js/react/components/files/FileManager/FileManager.module.scss
