+++ indico/web/client/js/react/components/files/FileManager/Uploads.jsx
