+++ indico/web/client/js/react/components/files/FileManager/__tests__/FileManager.spec.jsx
