+++ indico/web/client/js/react/components/files/FileManager/actions.js
