+++ indico/web/client/js/react/components/files/FileManager/reducer.js
