+++ indico/web/client/js/react/components/files/FileManager/selectors.js
