+++ indico/web/client/js/react/components/files/FileManager/util.js
